Use a separate chunk definition for each demo or text.
- In some versions of Wolfenstein 3D, not all text chunks are stored
continuously (T_HELPART is separated from T_ENDART* by demos).
While it may be possible to add support for definitions of multiple
ranges of such chunks, it is simpler to treat the texts like the
b800 and terminator texts i.e., special cases of misc. "extern" chunks.
In addition, Spear of Destiny has its own 11 palette chunks, which
currently need to be defined in a similar manner (using 11 lines).
Finally, this gives more flexibility for the choice of filenames.
- While there may be no known case where demos are not stored
continuously, they also belong to the category of "extern" chunks,
and so they now need separate definitions as well, just in case.