BREAKING CHANGE:
This modifies the expected format of masked 16x16 tiles in VGAGRAPH
files.
- Before: Four quarter-width 8-bpp planes of a single tile's data. One
of the 256 colors would internally be used for transparency (in-game).
- After: Two *linear* full-width 8-bpp planes, bitmap and mask.
This is the format originally used in an early Wolfenstein 3D alpha.

Note that the expected format of unmasked tiles remains unchanged
(and is still planar, rather than being linear).