PARTIALLY BREAKING CHANGE:
M16 tile no. 0 does not have to be imported as sparse anymore.
Basically, if the 16x16 masked tiles bitmap begins with a non-sparse
tile, then modid will import it as-is, instead of replacing it with
a sparse tile.

This was done in order to be able to faithfully create VGAGRAPH.WL1
from an early Wolfenstein 3D alpha, in which tile no. 128 mistakenly
got relocated to become tile no. 0.