Add hacks for support of exporting 98GRAPH.WL6.
Not a lot is known about the exact nature of 98GRAPH.WL6, so for now,
hacks are used. The hacks are enabled for 98GRAPH.EXT files, which are
otherwise processed in the same way as VGAGRAPH.EXT files.

A few notes:
- There are great chances the PC-98 port reads graphics data
from 98GRAPH.WL6 *and* VGAGRAPH.WL6 altogether,
while preferring the first file.
- Chunk no. 0 (bitmaps table) is NON-compressed, and has the WRONG
expanded size. It's also missing the dimensions of the last few
pictures (HUD face pics).
- The first 10 bitmaps, which are the HelpScreens pics, have their
own caveats. Basically, they're NON-compressed and the expanded size
is NOT embedded (assumed to be the raw size of chunk in 98GRAPH.WL6).
Furthermore, the graphics' pixels are arranged in simple LINEAR orders,
rather than the VGA-mode-suited planar order usually in use.