fin2bmp: Support platforms where long is not 32-bit
The fin2bmp code for handling external graphics screens in
Keen:Vorticons read the plane size as a "long". On some 64-bit platforms
(notably Unix-y ones), long is 64-bit, so not only is the wrong plane
size read, but also the first four bytes of data are skipped.

This change replaces long with uint32_t here, so that it will always be
32-bit. It also is now unsigned (which makes more sense). This will
still fail on big-endian setups, but none of the keen123 code supports
those, at least.